Remove vlc and change README.md
