+++ requirements.txt
@@ -1,2 +1,8 @@
+dbus-python==1.3.2
+decorator==5.1.1
+evento==1.0.2
+omxplayer-wrapper==0.3.3
+pathlib2==2.3.7.post1
+pkg-resources==0.0.0
 python-osc==1.8.0
-python-vlc==3.0.16120
+six==1.16.0
